+++ app/Http/Controllers/KategoryController.php
@@ -9,7 +9,7 @@ class KategoryController extends Controller
 {
     public function index()
     {
-        $datakategory = kategory::select('nama_kategory','id_kategory')->paginate(10);
+        $datakategory = kategory::select('nama_kategory','id_kategory')->get();
         return view('admin.kategory.indexkategory',compact('datakategory'));
     }
     public function store(Request $request)
@@ -36,8 +36,8 @@ class KategoryController extends Controller
     public function edit($kategory)
     {
         $data = kategory::findOrFail($kategory);
-      return view('admin.kategory.editkategory',compact('data'));
-
+        if($data == null){abort(404);}
+        return view('admin.kategory.editkategory',compact('data'));
     }
     public function update(Request $request,$kategory)
     { 
@@ -48,7 +48,7 @@ class KategoryController extends Controller
         ]);
 
         $data = kategory::findOrFail($kategory);
-
+        if($data == null){abort(404);}
         $data->update([
             'nama_kategory' => $request->nama_kategory,
         ]);
@@ -66,15 +66,9 @@ class KategoryController extends Controller
     public function destroy($kategory)
     {
         $data = kategory::findOrFail($kategory);
+        if($data == null){abort(404);}
         $data->delete();
         return redirect()->route('Kategory.index');
     }
-    public function search(Request $request)
-    {
-        $datakategory =kategory::select('nama_kategory','id_kategory')
-            ->where('nama_kategory','like','%'.$request->search.'%')
-            
-            ->paginate(10);
-            return view('admin.kategory.indexkategory',compact('datakategory'));
-    }
+   
 }
